+++ docs/plan/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 2026-09-07
 
+- docs/plan/characters/court-favorite.md
+- docs/plan/characters/military-war-record.md
 - docs/plan/characters/prestige.md
 - docs/plan/characters/skill-specializations.md